requirements.txt
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
daphne==2.4.1
Django==3.0.5
feedfinder2==0.0.4
feedparser==5.2.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jieba3k==0.35.1
lxml==4.5.0
mysqlclient==1.4.6
newspaper3k==0.2.8
nltk==3.4.5
Pillow==7.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-file==1.4.3
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tinysegmenter==0.3
tldextract==2.2.2
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
zope.interface==5.0.2
